refactor(knowledge): 移除冗余的警告卡片导入
在多个知识组件中移除了对 WarningCard 组件的导入，简化了代码结构，提升了可读性。相关文件包括 layout-basics.tsx、responsive.tsx、go_basics.tsx、go_concurrency.tsx、go_databases.tsx、go_testing.tsx 和 go_web_development.tsx。
