+++ src/components/knowledge/go/data/go_concurrency.tsx
@@ -1,9 +1,8 @@
-import { QuestionCard } from "../../../../base/knowledge_question_card";
-import { InfoCard } from "../../../../card/info_card";
-import { WarningCard } from "../../../../card/warning_card";
-import { SuccessCard } from "../../../../card/success_card";
-import { SecondaryCard } from "../../../../card/secondary_card";
-import { ExpandableCode } from "../../../../base/expandable_code";
+import { QuestionCard } from "../../../base/knowledge_question_card";
+import { InfoCard } from "../../../card/info_card";
+import { SuccessCard } from "../../../card/success_card";
+import { SecondaryCard } from "../../../card/secondary_card";
+import { ExpandableCode } from "../../../base/expandable_code";
 
 interface Props {
     id: string;
